third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/has-elementshadow-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
This test confirms that the number of elements having Shadow in a shadow subtree is correctly counted. See Bug 100922 also.

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
